backend/atomic-services/section/requirements.txt
flask
flask_sqlalchemy
marshmallow
python-dotenv
sqlalchemy
psycopg2-binary
flask-cors==6.0.0
flask-openapi3
pydantic

